backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
motor
pydantic
python-jose[cryptography]
python-dotenv
